Cleanup; restructure internal classes
