+++ src/Core/src/Internal/Common/Registry.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Spiral\Core\Internal;
+namespace Spiral\Core\Internal\Common;
 
 use Spiral\Core\Config;
