+++ src/Core/src/Internal/Invoker.php
@@ -7,6 +7,8 @@ namespace Spiral\Core\Internal;
 use Psr\Container\ContainerInterface;
 use Spiral\Core\Exception\Container\ContainerException;
 use Spiral\Core\Exception\Container\NotCallableException;
+use Spiral\Core\Internal\Common\DestructorTrait;
+use Spiral\Core\Internal\Common\Registry;
 use Spiral\Core\InvokerInterface;
 use Spiral\Core\ResolverInterface;
