+++ src/Core/src/Internal/Resolver/ResolvingState.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Spiral\Core\Internal;
+namespace Spiral\Core\Internal\Resolver;
 
 use ReflectionFunctionAbstract;
 use ReflectionParameter;
