+++ src/Core/src/Internal/Tracer/Trace.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace Spiral\Core\Internal;
+namespace Spiral\Core\Internal\Tracer;
 
-use ReflectionFunction;
 use Spiral\Core\Exception\Traits\ClosureRendererTrait;
 
 /**
