+++ src/Core/tests/Internal/BaseTest.php
@@ -9,8 +9,8 @@ use Spiral\Core\BinderInterface;
 use Spiral\Core\Config;
 use Spiral\Core\Container\InjectorInterface;
 use Spiral\Core\FactoryInterface;
+use Spiral\Core\Internal\Common\Registry;
 use Spiral\Core\Internal\Factory;
-use Spiral\Core\Internal\Registry;
 use Spiral\Core\Internal\Resolver;
 use Spiral\Core\Internal\State;
 use Spiral\Core\ResolverInterface;
